图部分路线的创建；Rail中id设为String，并修改相关函数，作为设置最大值为50

+++ HSRail/Common.h
@@ -15,7 +15,7 @@ typedef char String[18];
 
 typedef struct {
 	int day;
-	int hour;
+	float hour;
 }Date;//日期
 
 //羽丰牛逼！
